+++ release/graph/graph.component.js
@@ -25,13 +25,14 @@ var defaultDagreLayout = {
     ranksep: 100,
     acyclicer: undefined,
     ranker: 'network-simplex',
-    align: undefined,
+    align: undefined
 };
 var GraphComponent = /** @class */ (function (_super) {
     __extends(GraphComponent, _super);
     function GraphComponent() {
         var _this = _super !== null && _super.apply(this, arguments) || this;
         _this.nodes = [];
+        _this.dataLinks = [];
         _this.links = [];
         _this.activeEntries = [];
         _this.orientation = 'LR';
@@ -45,6 +46,9 @@ var GraphComponent = /** @class */ (function (_super) {
         _this.panOnZoom = true;
         _this.autoCenter = false;
         _this.dagreLayout = defaultDagreLayout;
+        _this.singleNodesPerLine = 1;
+        _this.multigraph = false;
+        _this.compound = false;
         _this.activate = new EventEmitter();
         _this.deactivate = new EventEmitter();
         _this.zoomChange = new EventEmitter();
@@ -261,6 +265,38 @@ var GraphComponent = /** @class */ (function (_super) {
                     node.width = Math.min(node.width, _this.nodeMinWidth);
             });
         }
+        if (this.singleNodesPerLine > 1) {
+            var singleNodesAmount = 0;
+            var currentDestinationNode = void 0;
+            var components = dagre.graphlib.alg.components(this.graph);
+            var _loop_1 = function (i) {
+                if (components[i].length > 1) {
+                    return "continue";
+                }
+                singleNodesAmount++;
+                singleNodesAmount = singleNodesAmount % this_1.singleNodesPerLine;
+                // Starting a new nodes row by adding a new temp node
+                if (singleNodesAmount === 1) {
+                    currentDestinationNode = {
+                        id: id(),
+                        width: 20,
+                        height: 30,
+                    };
+                    this_1.graph.setNode(currentDestinationNode.id, currentDestinationNode);
+                }
+                // Finding the single node element
+                var removedNodeId = components[i][0];
+                var currentSourceNode = this_1._nodes.find(function (n) { return n.id === removedNodeId; });
+                // Connecting it to the last node in the row.
+                this_1.graph.setEdge(currentSourceNode.id, currentDestinationNode.id);
+                // Update last node in the row
+                currentDestinationNode = currentSourceNode;
+            };
+            var this_1 = this;
+            for (var i = 0; i < components.length; i++) {
+                _loop_1(i);
+            }
+        }
         // Dagre to recalc the layout
         dagre.layout(this.graph);
         // Tranposes view options to the node
@@ -274,16 +310,22 @@ var GraphComponent = /** @class */ (function (_super) {
         });
         // Update the labels to the new positions
         var newLinks = [];
-        var _loop_1 = function (k) {
-            var l = this_1.graph._edgeLabels[k];
+        var _loop_2 = function (k) {
+            var l = this_2.graph._edgeLabels[k];
             var normKey = k.replace(/[^\w-]*/g, '');
-            var oldLink = this_1._oldLinks.find(function (ol) { return "" + ol.source + ol.target === normKey; });
+            var oldLink = this_2.multigraph ? this_2._oldLinks.find(function (ol) { return "" + ol.source + ol.target + ol.id === normKey; }) :
+                this_2._oldLinks.find(function (ol) { return "" + ol.source + ol.target === normKey; });
             if (!oldLink) {
-                oldLink = this_1._links.find(function (nl) { return "" + nl.source + nl.target === normKey; });
+                oldLink = this_2.multigraph ? this_2._links.find(function (nl) { return "" + nl.source + nl.target + nl.id === normKey; }) :
+                    this_2._links.find(function (nl) { return "" + nl.source + nl.target === normKey; });
+            }
+            // This is for supporting multiple single nodes in a row
+            if (!oldLink) {
+                return "continue";
             }
             oldLink.oldLine = oldLink.line;
             var points = l.points;
-            var line = this_1.generateLine(points);
+            var line = this_2.generateLine(points);
             var newLink = Object.assign({}, oldLink);
             newLink.line = line;
             newLink.points = points;
@@ -295,12 +337,12 @@ var GraphComponent = /** @class */ (function (_super) {
             if (!newLink.oldLine) {
                 newLink.oldLine = newLink.line;
             }
-            this_1.calcDominantBaseline(newLink);
+            this_2.calcDominantBaseline(newLink);
             newLinks.push(newLink);
         };
-        var this_1 = this;
+        var this_2 = this;
         for (var k in this.graph._edgeLabels) {
-            _loop_1(k);
+            _loop_2(k);
         }
         this._links = newLinks;
         // Map the old links for animations
@@ -322,6 +364,7 @@ var GraphComponent = /** @class */ (function (_super) {
             this.center();
         }
         requestAnimationFrame(function () { return _this.redrawLines(); });
+        this.redrawDataLinks();
         this.cd.markForCheck();
     };
     /**
@@ -344,13 +387,40 @@ var GraphComponent = /** @class */ (function (_super) {
                 .transition()
                 .duration(_animate ? 500 : 0)
                 .attr('d', l.line);
-            _this.handleLinkDataUIRedraw(linkEl, l);
             var textPathSelection = select(_this.chartElement.nativeElement).select("#" + l.id);
             textPathSelection
                 .attr('d', l.oldTextPath)
                 .transition()
                 .duration(_animate ? 500 : 0)
                 .attr('d', l.textPath);
+        });
+    };
+    GraphComponent.prototype.redrawDataLinks = function () {
+        var _this = this;
+        this.dataLinkElements.map(function (dataLinkEl) {
+            // Finding the datalink element
+            var datalink = _this._dataLinks.find(function (lin) { return lin.id === dataLinkEl.nativeElement.id; });
+            if (!datalink)
+                return;
+            // Finding the line that we want to place it on.
+            var line = _this._links.find(function (l) { return (l.source === datalink.source) && (l.target === datalink.target); });
+            if (!line)
+                return;
+            var dataLinkUI = select(dataLinkEl.nativeElement).select('.linkDataUI');
+            switch (line.points.length) {
+                case 1:
+                    break;
+                case 2:
+                    // Placing the link data template in the middle point between the start point and the end.
+                    var middleX = (line.points[0].x + line.points[1].x) / 2;
+                    var middleY = (line.points[0].y + line.points[1].y) / 2;
+                    dataLinkUI.attr('transform', "translate(" + middleX + ", " + middleY + ")");
+                    break;
+                default:
+                    // Placing the link data template in the middle point
+                    var middlePointIndex = Math.floor(line.points.length / 2);
+                    dataLinkUI.attr('transform', "translate(" + line.points[middlePointIndex].x + ", \n                                                  " + line.points[middlePointIndex].y + ")");
+            }
         });
     };
     /**
@@ -361,7 +431,11 @@ var GraphComponent = /** @class */ (function (_super) {
      */
     GraphComponent.prototype.createGraph = function () {
         var _this = this;
-        this.graph = new dagre.graphlib.Graph();
+        var graphParameters = {
+            multigraph: this.multigraph,
+            compound: this.compound
+        };
+        this.graph = new dagre.graphlib.Graph(graphParameters);
         this.graph.setGraph({
             rankdir: this.orientation,
             nodesep: this.dagreLayout.nodesep ? this.dagreLayout.nodesep : defaultDagreLayout.nodesep,
@@ -386,6 +460,9 @@ var GraphComponent = /** @class */ (function (_super) {
                 newLink.id = id();
             return newLink;
         });
+        this._dataLinks = this._links.map(function (l) {
+            return Object.assign({}, l);
+        });
         for (var _i = 0, _a = this._nodes; _i < _a.length; _i++) {
             var node = _a[_i];
             if (node.width && node.height) {
@@ -404,7 +481,8 @@ var GraphComponent = /** @class */ (function (_super) {
         // update dagre
         for (var _b = 0, _c = this._links; _b < _c.length; _b++) {
             var edge = _c[_b];
-            this.graph.setEdge(edge.source, edge.target);
+            this.multigraph ? this.graph.setEdge(edge.source, edge.target, edge, edge.id) :
+                this.graph.setEdge(edge.source, edge.target);
         }
         requestAnimationFrame(function () { return _this.draw(); });
     };
@@ -476,9 +554,9 @@ var GraphComponent = /** @class */ (function (_super) {
             point.y = mouseY;
             var svgPoint = point.matrixTransform(svgGroup.getScreenCTM().inverse());
             // Panzoom
-            this.pan(svgPoint.x, svgPoint.y);
+            this.pan(svgPoint.x, svgPoint.y, true);
             this.zoom(zoomFactor);
-            this.pan(-svgPoint.x, -svgPoint.y);
+            this.pan(-svgPoint.x, -svgPoint.y, true);
         }
         else {
             this.zoom(zoomFactor);
@@ -490,8 +568,9 @@ var GraphComponent = /** @class */ (function (_super) {
      * @param x
      * @param y
      */
-    GraphComponent.prototype.pan = function (x, y) {
-        var zoomLevel = this.zoomLevel;
+    GraphComponent.prototype.pan = function (x, y, ignoreZoomLevel) {
+        if (ignoreZoomLevel === void 0) { ignoreZoomLevel = false; }
+        var zoomLevel = ignoreZoomLevel ? 1 : this.zoomLevel;
         this.transformationMatrix = transform(this.transformationMatrix, translate(x / zoomLevel, y / zoomLevel));
         this.updateTransform();
     };
@@ -556,10 +635,10 @@ var GraphComponent = /** @class */ (function (_super) {
         var x = node.x - node.width / 2;
         var y = node.y - node.height / 2;
         node.options.transform = "translate(" + x + ", " + y + ")";
-        var _loop_2 = function (link) {
+        var _loop_3 = function (link) {
             if (link.target === node.id || link.source === node.id) {
-                var sourceNode = this_2._nodes.find(function (n) { return n.id === link.source; });
-                var targetNode = this_2._nodes.find(function (n) { return n.id === link.target; });
+                var sourceNode = this_3._nodes.find(function (n) { return n.id === link.source; });
+                var targetNode = this_3._nodes.find(function (n) { return n.id === link.target; });
                 // determine new arrow position
                 var dir = sourceNode.y <= targetNode.y ? -1 : 1;
                 var startingPoint = {
@@ -572,18 +651,19 @@ var GraphComponent = /** @class */ (function (_super) {
                 };
                 // generate new points
                 link.points = [startingPoint, endingPoint];
-                var line = this_2.generateLine(link.points);
-                this_2.calcDominantBaseline(link);
+                var line = this_3.generateLine(link.points);
+                this_3.calcDominantBaseline(link);
                 link.oldLine = link.line;
                 link.line = line;
             }
         };
-        var this_2 = this;
+        var this_3 = this;
         for (var _i = 0, _a = this._links; _i < _a.length; _i++) {
             var link = _a[_i];
-            _loop_2(link);
+            _loop_3(link);
         }
         this.redrawLines(false);
+        this.redrawDataLinks();
     };
     /**
      * Update the entire view for the new pan position
@@ -801,26 +881,6 @@ var GraphComponent = /** @class */ (function (_super) {
         }
         this.panTo(node.x, node.y);
     };
-    GraphComponent.prototype.handleLinkDataUIRedraw = function (linkEl, l) {
-        var linkCenter = select(linkEl.nativeElement).select('.linkDataUI');
-        if (!linkCenter || !linkEl || !l || !l.points) {
-            return;
-        }
-        switch (l.points.length) {
-            case 1:
-                break;
-            case 2:
-                // Placing the link data template in the middle point between the start point and the end.
-                var middleX = (l.points[0].x + l.points[1].x) / 2;
-                var middleY = (l.points[0].y + l.points[1].y) / 2;
-                linkCenter.attr('transform', "translate(" + middleX + ", " + middleY + ")");
-                break;
-            default:
-                // Placing the link data template in the middle point
-                var middlePointIndex = Math.floor(l.points.length / 2);
-                linkCenter.attr('transform', "translate(" + l.points[middlePointIndex].x + ", " + l.points[middlePointIndex].y + ")");
-        }
-    };
     GraphComponent.decorators = [
         { type: Component, args: [{
                     selector: 'ngx-graph',
@@ -828,12 +888,13 @@ var GraphComponent = /** @class */ (function (_super) {
                     encapsulation: ViewEncapsulation.None,
                     changeDetection: ChangeDetectionStrategy.OnPush,
                     animations: [trigger('link', [ngTransition('* => *', [animate(500, style({ transform: '*' }))])])],
-                    template: "\n    <ngx-charts-chart\n      [view]=\"[width, height]\"\n      [showLegend]=\"legend\"\n      [legendOptions]=\"legendOptions\"\n      (legendLabelClick)=\"onClick($event)\"\n      (legendLabelActivate)=\"onActivate($event)\"\n      (legendLabelDeactivate)=\"onDeactivate($event)\"\n      mouseWheel\n      (mouseWheelUp)=\"onZoom($event, 'in')\"\n      (mouseWheelDown)=\"onZoom($event, 'out')\"\n    >\n      <svg:g\n        *ngIf=\"initialized\"\n        [attr.transform]=\"transform\"\n        (touchstart)=\"onTouchStart($event)\"\n        (touchend)=\"onTouchEnd($event)\"\n        class=\"graph chart\"\n      >\n        <defs>\n          <ng-template *ngIf=\"defsTemplate\" [ngTemplateOutlet]=\"defsTemplate\"></ng-template>\n          <svg:path\n            class=\"text-path\"\n            *ngFor=\"let link of _links\"\n            [attr.d]=\"link.textPath\"\n            [attr.id]=\"link.id\"\n          ></svg:path>\n        </defs>\n        <svg:rect\n          class=\"panning-rect\"\n          [attr.width]=\"dims.width * 100\"\n          [attr.height]=\"dims.height * 100\"\n          [attr.transform]=\"'translate(' + (-dims.width || 0) * 50 + ',' + (-dims.height || 0) * 50 + ')'\"\n          (mousedown)=\"isPanning = true\"\n        />\n        <svg:g class=\"links\">\n          <svg:g *ngFor=\"let link of _links; trackBy: trackLinkBy\" class=\"link-group\" #linkElement [id]=\"link.id\">\n            <ng-template\n              *ngIf=\"linkTemplate\"\n              [ngTemplateOutlet]=\"linkTemplate\"\n              [ngTemplateOutletContext]=\"{ $implicit: link }\"\n            ></ng-template>\n            <ng-template\n              *ngIf=\"linkDataTemplate\"\n              [ngTemplateOutlet]=\"linkDataTemplate\"\n              [ngTemplateOutletContext]=\"{ $implicit: link }\"\n            ></ng-template>\n            <svg:path *ngIf=\"!linkTemplate\" class=\"edge\" [attr.d]=\"link.line\" />\n          </svg:g>\n        </svg:g>\n        <svg:g class=\"nodes\">\n          <svg:g\n            *ngFor=\"let node of _nodes; trackBy: trackNodeBy\"\n            class=\"node-group\"\n            #nodeElement\n            [id]=\"node.id\"\n            [attr.transform]=\"node.options.transform\"\n            (click)=\"onClick(node)\"\n            (mousedown)=\"onNodeMouseDown($event, node)\"\n          >\n            <ng-template\n              *ngIf=\"nodeTemplate\"\n              [ngTemplateOutlet]=\"nodeTemplate\"\n              [ngTemplateOutletContext]=\"{ $implicit: node }\"\n            ></ng-template>\n            <svg:circle\n              *ngIf=\"!nodeTemplate\"\n              r=\"10\"\n              [attr.cx]=\"node.width / 2\"\n              [attr.cy]=\"node.height / 2\"\n              [attr.fill]=\"node.options.color\"\n            />\n          </svg:g>\n        </svg:g>\n      </svg:g>\n    </ngx-charts-chart>\n  "
+                    template: "\n    <ngx-charts-chart\n      [view]=\"[width, height]\"\n      [showLegend]=\"legend\"\n      [legendOptions]=\"legendOptions\"\n      (legendLabelClick)=\"onClick($event)\"\n      (legendLabelActivate)=\"onActivate($event)\"\n      (legendLabelDeactivate)=\"onDeactivate($event)\"\n      mouseWheel\n      (mouseWheelUp)=\"onZoom($event, 'in')\"\n      (mouseWheelDown)=\"onZoom($event, 'out')\"\n    >\n      <svg:g\n        *ngIf=\"initialized\"\n        [attr.transform]=\"transform\"\n        (touchstart)=\"onTouchStart($event)\"\n        (touchend)=\"onTouchEnd($event)\"\n        class=\"graph chart\"\n      >\n        <defs>\n          <ng-template *ngIf=\"defsTemplate\" [ngTemplateOutlet]=\"defsTemplate\"></ng-template>\n          <svg:path\n            class=\"text-path\"\n            *ngFor=\"let link of _links\"\n            [attr.d]=\"link.textPath\"\n            [attr.id]=\"link.id\"\n          ></svg:path>\n        </defs>\n        <svg:rect\n          class=\"panning-rect\"\n          [attr.width]=\"dims.width * 100\"\n          [attr.height]=\"dims.height * 100\"\n          [attr.transform]=\"'translate(' + (-dims.width || 0) * 50 + ',' + (-dims.height || 0) * 50 + ')'\"\n          (mousedown)=\"isPanning = true\"\n        />\n        <svg:g class=\"links\">\n          <svg:g *ngFor=\"let link of _links; trackBy: trackLinkBy\" class=\"link-group\" #linkElement [id]=\"link.id\">\n            <ng-template\n              *ngIf=\"linkTemplate\"\n              [ngTemplateOutlet]=\"linkTemplate\"\n              [ngTemplateOutletContext]=\"{ $implicit: link }\"\n            ></ng-template>\n            <svg:path *ngIf=\"!linkTemplate\" class=\"edge\" [attr.d]=\"link.line\" />\n          </svg:g>\n        </svg:g>\n        <svg:g class=\"datalinks\" *ngIf=\"linkDataTemplate\">\n          <svg:g *ngFor=\"let datalink of _dataLinks; trackBy: trackNodeBy\" class=\"datalink-group\" \n                  #datalinkElement [id]=\"datalink.id\">\n            <ng-template\n              [ngTemplateOutlet]=\"linkDataTemplate\"\n              [ngTemplateOutletContext]=\"{ $implicit: datalink }\"\n            ></ng-template>\n          </svg:g>\n        </svg:g>\n        <svg:g class=\"nodes\">\n          <svg:g\n            *ngFor=\"let node of _nodes; trackBy: trackNodeBy\"\n            class=\"node-group\"\n            #nodeElement\n            [id]=\"node.id\"\n            [attr.transform]=\"node.options.transform\"\n            (click)=\"onClick(node)\"\n            (mousedown)=\"onNodeMouseDown($event, node)\"\n          >\n            <ng-template\n              *ngIf=\"nodeTemplate\"\n              [ngTemplateOutlet]=\"nodeTemplate\"\n              [ngTemplateOutletContext]=\"{ $implicit: node }\"\n            ></ng-template>\n            <svg:circle\n              *ngIf=\"!nodeTemplate\"\n              r=\"10\"\n              [attr.cx]=\"node.width / 2\"\n              [attr.cy]=\"node.height / 2\"\n              [attr.fill]=\"node.options.color\"\n            />\n          </svg:g>\n        </svg:g>\n      </svg:g>\n    </ngx-charts-chart>\n  "
                 },] },
     ];
     GraphComponent.propDecorators = {
         legend: [{ type: Input }],
         nodes: [{ type: Input }],
+        dataLinks: [{ type: Input }],
         links: [{ type: Input }],
         activeEntries: [{ type: Input }],
         orientation: [{ type: Input }],
@@ -858,6 +919,9 @@ var GraphComponent = /** @class */ (function (_super) {
         zoomToFit$: [{ type: Input }],
         zoomToNode$: [{ type: Input }],
         dagreLayout: [{ type: Input }],
+        singleNodesPerLine: [{ type: Input }],
+        multigraph: [{ type: Input }],
+        compound: [{ type: Input }],
         activate: [{ type: Output }],
         deactivate: [{ type: Output }],
         zoomChange: [{ type: Output }],
@@ -868,6 +932,7 @@ var GraphComponent = /** @class */ (function (_super) {
         chart: [{ type: ViewChild, args: [ChartComponent, { read: ElementRef },] }],
         nodeElements: [{ type: ViewChildren, args: ['nodeElement',] }],
         linkElements: [{ type: ViewChildren, args: ['linkElement',] }],
+        dataLinkElements: [{ type: ViewChildren, args: ['datalinkElement',] }],
         groupResultsBy: [{ type: Input }],
         zoomLevel: [{ type: Input, args: ['zoomLevel',] }],
         panOffsetX: [{ type: Input, args: ['panOffsetX',] }],